Modified the move_into_check? tests.
The tests as previously written were passing even though the King's
movement now fails in the browser. The problem was that some of the
pieces instead of just returning false in their valid_move? are instead
raising errors when they call valid_move on a square they cannot move
to. The previous tests destroyed all pieces on the board, so these
deficient valid_move methods were not being called.