Wrote move_into_check? method on Piece model
Moved the move_into_check method to the Piece model. It works by
removing any piece that was attacked from the board, then moving the
called piece to the given coordinates, then calling game.check?
If the player is now in check, everything is moved back to it's original
position and true is returned.

I need to write some more tests, then set the metod so it works for
enpassants, then call it in the move method.