Piece#move_into_check? now works with en passants
Wrote tests for move_into_check method. Got the enpassant working.
The method now places all the pieces back regardless of whether or not
it passes. I did this so that it will work for castling. Though it will
cause a lot more writing to the database, so if you think of a way to do
it better, I'm all ears.