Calling Piece#move_into_check? in the move method.
Also King#can_castle only returns true if the squares that the king
moves through during the castle are not in check.

I added a king in some of the specs. When a piece moves it checks that
it's king is not being placed in check. If there is no king of the same
color in the game an ActiveRecord error is raised.