Added wins and losses columns to player model.
Modified Game#player_in_checkmate method to increment the appropiate
player's wins and losses on finding a checkmate. I'm concerned that
because this method is called from the view it is possible for it to be
called multiple times, i.e. from a page refresh, or just from the two
views of the players. I'm considering adding a column to the game table
to prevent a checkmate from being called multiple times.d