Created a can move to method for pieces.
I created a method can_move_to on the pieces model. This method is much
more simple than the already existant valid_move method. It merely
checks that the proposed move is in an array of coordinates that is
supplied by the particular piece that the piece could move to before
checking for other pieces on the board.

I am trying to remove responsiblities from the piece model. A single
piece should know where it can move, but does not need to know where all
the other pieces are on the board.