Added possible moves method to pawn model.
This implementation returns true for any move that could be a correct
move for a pawn without being concerned with what other pieces are
around. Later we will need to find a way to determine that a move is
attacking another piece, but it isn't the pawns responsibility to know
where other pieces are, so this will have to be implemented in the Move
class.